+++ graphics/r_shaders.cpp
@@ -1,0 +1,7 @@
+//project fresa, 2017-2022
+//by jose pazos perez
+//licensed under GPLv3
+
+#include "r_shaders.h"
+
+using namespace Fresa;